refactored toolBox and toolUpload functionality out of app and into lib
