+++ lib/src/editor/index.ts
@@ -4,6 +4,7 @@ export * from './EntityList'
 export * from './EntityViewer'
 export * from './PropertyEditor'
 export * from './SelectProperty'
+export * from './ToolUpload'
 export * from './VertexCreateDialog'
 export * from './VertexSchemaSelect'
 export * from './VertexSelect'
